Moved ProductController to new package
